combined tui and cli together, one Mir binary to rule them all
