+++ docs/idea.md
@@ -1,0 +1,5 @@
+# Idea
+
+default cli for service, add a command that
+create the config file with the default configuration
+instead of copy pasting manually the cli output
