+++ makefile
@@ -1,21 +1,29 @@
 .PHONY: api seed build
 
+# code generation
 api:
 	buf generate --config ./api/buf.gen.yaml
 
+# scripts
 seed: build
 	./scripts/seed.sh
 
 clean_db:
 	./scripts/clean_db.sh
 
+# builds
 build:
-	go build -o bin/mir cmds/cli/main.go
 	go build -o bin/tui cmds/tui/main.go
+	go build -o bin/mir cmds/mir/main.go
 	go build -o bin/core cmds/core/main.go
 
-tui-log:
-	tail ~/.config/mir/cli.log -f
+mir-log:
+	tail ~/.config/mir/mir.log -f
 
+# install
+install-mir:
+	sudo cp bin/mir /usr/local/bin/mir
+
+# air
 air-core:
 	air -c .air/core.toml
